requirements.txt
keras
pillow
numpy
